requirements.txt
fastapi
uvicorn
sqlalchemy
pymysql
python-dotenv
boto3
pymupdf
sentence-transformers
faiss-cpu
chromadb
langchain-community